refactor: builder services to extensions
