+++ Backend/Extensions/ApplicationServiceExtensions.cs
@@ -1,0 +1,26 @@
+using Backend.Data;
+using Backend.Interfaces;
+using Backend.Services;
+using Microsoft.EntityFrameworkCore;
+
+namespace Backend.Extensions;
+
+// static allows us to use the methods inside this class without creating a new instance
+public static class ApplicationServiceExtensions
+{
+  public static IServiceCollection AddApplicationServices(
+    this IServiceCollection services,
+    IConfiguration config
+  )
+  {
+    services.AddControllers();
+    services.AddDbContext<DataContext>(opt =>
+    {
+      opt.UseSqlite(config.GetConnectionString("DefaultConnection"));
+    });
+    services.AddCors();
+    services.AddScoped<ITokenService, TokenService>(); // create once per http request
+
+    return services;
+  }
+}
