+++ Backend/Program.cs
@@ -1,36 +1,10 @@
-using System.Text;
-using Backend.Data;
-using Backend.Interfaces;
-using Backend.Services;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
+using Backend.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-
-builder.Services.AddControllers();
-builder.Services.AddDbContext<DataContext>(opt =>
-{
-    opt.UseSqlite(builder.Configuration.GetConnectionString("DefaultConnection"));
-});
-builder.Services.AddCors();
-builder.Services.AddScoped<ITokenService, TokenService>(); // create once per http request
-builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-    .AddJwtBearer(options =>
-    {
-        var tokenKey = builder.Configuration["TokenKey"] ?? throw new Exception("TokenKey not found");
-        var tokenKeyByteArray = Encoding.UTF8.GetBytes(tokenKey);
-
-        options.TokenValidationParameters = new TokenValidationParameters
-        {
-            ValidateIssuerSigningKey = true, // must, otherwise it will accept any token, unsigned or not
-            IssuerSigningKey = new SymmetricSecurityKey(tokenKeyByteArray),
-            ValidateIssuer = false, // not passing in, so not needed for now
-            ValidateAudience = false // not passing in, so not needed for now
-        };
-    });
+builder.Services.AddApplicationServices(builder.Configuration);
+builder.Services.AddIdentityServices(builder.Configuration);
 
 var app = builder.Build();
